requirements.txt
Django==2.0.5
django-rest-framework==0.1.0
djangorestframework==3.8.2
psycopg2-binary==2.7.4
python-chess==0.23.4
pytz==2018.4
